refactor(scripts): file-organization sweep over oxlint
Each plugin entrypoint keeps only its definePlugin call and header where .oxlintrc.json loads it; every rule, helper and message moves to one file each under services/oxlint/<plugin>, and the notify rule's frame and binding shapes become model files. The fixture harness resolves the entrypoint directory and the oxlint binary from REPOSITORY_ROOT rather than counting its own depth. The oxlint skill's authoring page and the coderabbit cut-verification recipe carry the split.
