+++ .agents/skills/coderabbit/references/window-composition.md
@@ -35,7 +35,7 @@ flowchart TD
 
 **Re-splitting the unpushed range is free; a pushed commit is never rewritten.** With a clean tree, `git reset --soft origin/<branch>` puts every unpushed change back in the index and `git add <paths>` / `git commit` re-cuts it in the order the window needs; the reordering recipe below is the same operation for whole commits. Both stop at `origin/<branch>` — that ref is the reviewed frontier and other sessions' pushes, and a rewrite past it is the checkpoint desync the last section describes.
 
-**Verify the cut, not only the head.** The checks ran against the tip, which includes the held tail; the cut's tree is that tip minus the tail, so what the tail touches is re-run against the cut — the package suites whose snapshots it moved, the docs suite whose paths it fixed, and lint with the plugin at the cut (`git show <cut>:scripts/src/oxlint/<plugin>.ts` into a scratch file beside the others, a scratch config at the repo root naming it, deleted after).
+**Verify the cut, not only the head.** The checks ran against the tip, which includes the held tail; the cut's tree is that tip minus the tail, so what the tail touches is re-run against the cut — the package suites whose snapshots it moved, the docs suite whose paths it fixed, and lint with the plugin at the cut (`git worktree add <scratch> <cut>`, a scratch config at the repo root naming that worktree's `scripts/src/oxlint/<plugin>.ts`, both removed after — the entrypoint alone is not the plugin, since its rules resolve through `#src/services/oxlint/<plugin>/`).
 
 **Fill to the cap.** A slot costs an hour whether it reads 20 files or 99, so the prefix is the largest one under the cap, and a prefix that lands at 99 is pushed rather than held at 84. The ~90 figure the skill aims at is what a prefix usually lands on, never a reason to hold one that fits. Other sessions push the same branch, so the count is re-read after a fetch and the push goes out only while the remote tip is still the one it was measured against.
