+++ .agents/skills/oxlint/references/custom-js-plugins.md
@@ -35,9 +35,9 @@ A plugin that passes every gate is still re-read against them when its scope is 
 
 ## Writing one
 
-- Plugins are **TypeScript** files under `scripts/src/oxlint/` (one rule-set per file), so the root `tsc` typecheck covers them and oxlint loads them directly via Node type-stripping. Author them with `@oxlint/plugins`: `definePlugin`/`defineRule` (their sole purpose is inference — visitor handler params like `AwaitExpression(node)` type themselves, so **never annotate them inline**), and the `ESTree` namespace / `Context`/`Plugin` types for standalone helpers. There is no node-type enum — `node.type === "CallExpression"` literals are the discriminants, checked against `ESTree` so a typo won't compile. `@oxlint/plugins` and `oxlint` are catalogued as a caret pair and bumped together by the dependency-updates sweep — the repo tracks latest rather than pinning, so keep the two entries in step there.
+- A plugin is a **TypeScript** entrypoint at `scripts/src/oxlint/<plugin>.ts` — the `definePlugin` call and the header, nothing else — with each rule and helper in its own file under `scripts/src/services/oxlint/<plugin>/` and its types under `scripts/src/models/oxlint/<plugin>/` (`file-organization` skill), so the root `tsc` typecheck covers them and oxlint loads the entrypoint directly via Node type-stripping. Author them with `@oxlint/plugins`: `definePlugin`/`defineRule` (their sole purpose is inference — visitor handler params like `AwaitExpression(node)` type themselves, so **never annotate them inline**), and the `ESTree` namespace / `Context`/`Plugin` types for standalone helpers. There is no node-type enum — `node.type === "CallExpression"` literals are the discriminants, checked against `ESTree` so a typo won't compile. `@oxlint/plugins` and `oxlint` are catalogued as a caret pair and bumped together by the dependency-updates sweep — the repo tracks latest rather than pinning, so keep the two entries in step there.
 - Reference the `.ts` by path in the root `jsPlugins` array; enable the rule under `rules` (or a scoped `overrides` entry) as `<meta.name>/<rule>`.
-- **Every plugin is paired with a fixture suite that drives the real binary** — a rule with no suite is a rule nothing proves still fires. `setupPluginSuite` is the harness that runs one over its fixtures, in a single `beforeAll` pass shared by every fixture in the file. `scripts/src/oxlint/` is the set; each file's header states which convention it enforces, which docs page or skill owns that convention, and which of the three permitted lists its scope is — the construct's domain, the mechanism's sites, or a library's vocabulary.
+- **Every plugin is paired with a fixture suite that drives the real binary** — a rule with no suite is a rule nothing proves still fires. `setupPluginSuite` is the harness that runs one over its fixtures, in a single `beforeAll` pass shared by every fixture in the file. `scripts/src/oxlint/` is the set; each entrypoint's header states which convention it enforces, which docs page or skill owns that convention, and which of the three permitted lists its scope is — the construct's domain, the mechanism's sites, or a library's vocabulary.
 - **A rule is on for the whole tree unless the construct it reads only exists in part of it**, and the violation list an unscoped run reports is cleared in the change that lands the rule — never scoped to the clean paths as a rollout stage, since nobody comes back to widen it and the rest of the tree reads as covered (the `sweeps` skill). **Scope with `overrides`** only when a rule genuinely applies to part of the tree — e.g. `persistThenNotify.ts` (the `apps/web/content/docs/architecture/persist-then-notify.md` enforcer) is scoped to `apps/web/server/**/*.ts`, because an `EventEmitter.emit` only means "realtime notify" in server mutations; client emitters (the Phaser game bus) are unrelated. `overrides` objects reject unknown keys — no `"//"` comment field; document intent in the plugin file's header instead.
 - The plugin runs in oxlint's single root pass, so it's fast enough to stay always-on. **The JS plugin API is alpha and not subject to semver** — re-verify every plugin after an `oxlint` bump (plant a violation and confirm it still fires); a contract change silently drops the rule while CI stays green.
 - The plugin file is itself linted by the repo's own oxlint+eslint pass (it lives under `scripts/`), so it must satisfy every repo convention — no `void` operator, sorted `Set`s (`perfectionist/sort-sets`), capitalized comments, comments on their own line.
