+++ scripts/src/oxlint/passThroughHelper.ts
@@ -1,6 +1,7 @@
-import type { ESTree, Plugin } from "@oxlint/plugins";
+import type { Plugin } from "@oxlint/plugins";
 
-import { definePlugin, defineRule } from "@oxlint/plugins";
+import { noForwardingWrapper } from "#src/services/oxlint/passThroughHelper/noForwardingWrapper";
+import { definePlugin } from "@oxlint/plugins";
 // An oxlint JS plugin enforcing the "an extraction earns its existence" rule (file-organization/SKILL.md).
 //
 // An exported arrow whose whole body is one call passing exactly its own parameters, in order, adding nothing,
@@ -22,90 +23,9 @@ import { definePlugin, defineRule } from "@oxlint/plugins";
 //
 // It is off for `*.test.ts`/`*.bench.ts` (root .oxlintrc.json, which takes no comments): a colocated module
 // Double exists precisely to mirror the real signature, so forwarding is the whole point there.
-const MESSAGE =
-  "Forwarding wrapper: every argument is still hand-written at the call site, so this only renames the callee. Inline it, or give it something to absorb — a constant, a narrowing, or a definition several call sites must agree on.";
-const UPPER_SNAKE_REGEX = /^[A-Z0-9_]+$/u;
-// The name a parameter binds, or undefined for a pattern/default — both mean the wrapper is doing more than
-// Forwarding, so an undefined here stops the check rather than failing it.
-const getParameterName = (parameter: ESTree.Node): string | undefined => {
-  if (parameter.type === "Identifier") return parameter.name;
-  else if (parameter.type === "RestElement" && parameter.argument.type === "Identifier") return parameter.argument.name;
-  else return undefined;
-};
-// The name an argument forwards, or undefined for anything the caller did not simply hand over.
-const getArgumentName = (argument: ESTree.Node): string | undefined => {
-  if (argument.type === "Identifier") return argument.name;
-  else if (argument.type === "SpreadElement" && argument.argument.type === "Identifier") return argument.argument.name;
-  else return undefined;
-};
-// The receiver a member call dispatches on, when it is a bare identifier: `client.deleteEntity(...)` -> `client`.
-const getReceiverName = (expression: ESTree.CallExpression | ESTree.NewExpression): string | undefined => {
-  if (expression.type !== "CallExpression" || expression.callee.type !== "MemberExpression") return undefined;
-  const { object } = expression.callee;
-  return object.type === "Identifier" ? object.name : undefined;
-};
-
-const rule = defineRule({
-  create(context) {
-    const checkIsForwardingCall = (
-      parameterNames: string[],
-      expression: ESTree.CallExpression | ESTree.NewExpression,
-    ): boolean => {
-      const receiverName = getReceiverName(expression);
-      // A member call is only a forward when the receiver came from the caller too; on anything else the wrapper
-      // Is supplying the object, and an UPPER_SNAKE one is supplying a constant outright
-      if (expression.type === "CallExpression" && expression.callee.type === "MemberExpression") {
-        if (receiverName === undefined || UPPER_SNAKE_REGEX.test(receiverName)) return false;
-      } else if (expression.callee.type !== "Identifier") return false;
-      const expectedNames = receiverName === parameterNames[0] ? parameterNames.slice(1) : parameterNames;
-      const argumentNames = expression.arguments.map((argument) => getArgumentName(argument));
-      if (argumentNames.length !== expectedNames.length) return false;
-      return argumentNames.every((argumentName, index) => argumentName === expectedNames[index]);
-    };
-    // A read of what a forward would have returned is the same rename one step further out:
-    // `() => useVTheme().global` and `(a) => a.b` both hand back a property the caller could have reached
-    // Itself.
-    const checkIsForwardingRead = (parameterNames: string[], expression: ESTree.MemberExpression): boolean => {
-      if (expression.computed) return false;
-      const { object } = expression;
-      if (object.type === "Identifier") return parameterNames.includes(object.name);
-      else if (object.type === "CallExpression" || object.type === "NewExpression")
-        return checkIsForwardingCall(parameterNames, object);
-      else return false;
-    };
-    // An exported arrow reaches the surface either named or as the module's default, and the two shapes forward
-    // Identically, so both visitors hand their arrow here.
-    const reportIfForwarding = (arrow: ESTree.ArrowFunctionExpression): void => {
-      const parameterNames = arrow.params
-        .map((parameter) => getParameterName(parameter))
-        .filter((parameterName) => parameterName !== undefined);
-      if (parameterNames.length !== arrow.params.length) return;
-      // `async (a) => await f(a)` forwards exactly as its sync twin does
-      const body = arrow.body.type === "AwaitExpression" ? arrow.body.argument : arrow.body;
-      const isForwarding =
-        body.type === "MemberExpression"
-          ? checkIsForwardingRead(parameterNames, body)
-          : (body.type === "CallExpression" || body.type === "NewExpression") &&
-            checkIsForwardingCall(parameterNames, body);
-      if (isForwarding) context.report({ message: MESSAGE, node: arrow });
-    };
-    return {
-      ExportDefaultDeclaration(node) {
-        if (node.declaration.type === "ArrowFunctionExpression") reportIfForwarding(node.declaration);
-      },
-      ExportNamedDeclaration(node) {
-        if (node.declaration?.type !== "VariableDeclaration") return;
-        for (const { init } of node.declaration.declarations)
-          if (init?.type === "ArrowFunctionExpression") reportIfForwarding(init);
-      },
-    };
-  },
-  meta: { type: "suggestion" },
-});
-
 const plugin: Plugin = definePlugin({
   meta: { name: "pass-through-helper" },
-  rules: { "no-forwarding-wrapper": rule },
+  rules: { "no-forwarding-wrapper": noForwardingWrapper },
 });
 
 export default plugin;
