+++ scripts/src/oxlint/persistThenNotify.test.ts
@@ -1,4 +1,4 @@
-import { setupPluginSuite } from "#src/oxlint/setupPluginSuite.test";
+import { setupPluginSuite } from "#src/services/oxlint/setupPluginSuite.test";
 import { describe, expect, test } from "vitest";
 
 const RULE = "persist-then-notify/no-unhandled-effect-after-emit";
