+++ scripts/src/services/oxlint/errorAlert/constants.ts
@@ -1,0 +1,2 @@
+export const MESSAGE =
+  "A rejection reaches the user through createErrorAlert(error), which asks the error link first — alerting error.message directly stacks a second toast on a failure the link already showed.";
