+++ scripts/src/services/oxlint/literalUnion/constants.ts
@@ -1,0 +1,2 @@
+export const MESSAGE =
+  'A union of string literals is an enum — declare `enum Foo { Bar = "bar" }` in its own model file and reference it. See the typescript skill.';
