+++ scripts/src/services/oxlint/trpcProcedure/ErrorGuardMap.ts
@@ -1,0 +1,5 @@
+export const ErrorGuardMap: Record<string, string> = {
+  ForbiddenError: "getForbiddenError",
+  InvalidOperationError: "getInvalidOperationError",
+  NotFoundError: "getNotFoundError",
+};
